Changed "== null" tests to "is null" to avoid incorrect results with wrong overrides of object.Equals() (fixes #10)
When using == null, it's possible that the code is lying by overriding the Equals(obj) method and returning true when obj is null. Or even worse, when there's a bug in the overriden Equals(obj) that causes it to throw for example a NullReferenceException. We want Pitcher to throw an exception when something is null, not when something might possible in some cases sometimes maybe null.